requirements.txt
PyQt6>=6.4.0
paho-mqtt>=2.0.0
pytest==8.3.4
pyinstaller>=6.5.0